requirements.txt
# 웹 스크래핑
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
playwright==1.40.0

# 데이터 처리
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2

# 데이터베이스
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# API 서버
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# 비동기 처리
aiohttp==3.9.0
asyncio==3.4.3

# 유틸리티
python-dotenv==1.0.0
loguru==0.7.2
schedule==1.2.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0